school_management/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.0
Pillow>=10.0
reportlab>=4.0
openpyxl>=3.1
